parent-child support for all event

+++ api/core/rag/index_processor/processor/parent_child_index_processor.py
@@ -104,10 +104,13 @@ class ParentChildIndexProcessor(BaseIndexProcessor):
                 )
                 child_node_ids = [child_node_id[0] for child_node_id in child_node_ids]
                 vector.delete_by_ids(child_node_ids)
-                db.session.query(ChildChunk).filter(ChildChunk.dataset_id == dataset.id, ChildChunk.index_node_id.in_(child_node_ids)).delete()
-                db.session.commit()
             else:
                 vector.delete()
+            
+            delete_child_chunks = kwargs.get("delete_child_chunks") or False
+            if delete_child_chunks: 
+                db.session.query(ChildChunk).filter(ChildChunk.dataset_id == dataset.id, ChildChunk.index_node_id.in_(child_node_ids)).delete()
+                db.session.commit()
 
     def retrieve(
         self,
